initial draft: add/extendable-transformers
