+++ includes/class-scheduler.php
@@ -4,7 +4,6 @@ namespace Activitypub;
 
 use Activitypub\Collection\Users;
 use Activitypub\Collection\Followers;
-use Activitypub\Transformer\Post;
 
 /**
  * ActivityPub Scheduler Class
