+++ templates/post-json.php
@@ -2,8 +2,10 @@
 // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 $post = \get_post();
 
-$object = new \Activitypub\Transformer\Post( $post );
-$json = \array_merge( array( '@context' => \Activitypub\get_context() ), $object->to_object()->to_array() );
+$transformer = \Activitypub\Transformers_Manager::instance()->get_transformer( $post );
+$transformer->set_wp_post( $post );
+
+$json = \array_merge( array( '@context' => \Activitypub\get_context() ), $transformer->to_object()->to_array() );
 
 // filter output
 $json = \apply_filters( 'activitypub_json_post_array', $json );
