+++ tests/test-class-activitypub-post.php
@@ -10,13 +10,13 @@ class Test_Activitypub_Post extends WP_UnitTestCase {
 
 		$permalink = \get_permalink( $post );
 
-		$activitypub_post = \Activitypub\Transformer\Post::transform( get_post( $post ) )->to_object();
+		$activitypub_post = \Activitypub\Transformers_Manager::get_transformer( get_post( $post ) )->transform( get_post( $post ) )->to_object();
 
 		$this->assertEquals( $permalink, $activitypub_post->get_id() );
 
 		\wp_trash_post( $post );
 
-		$activitypub_post = \Activitypub\Transformer\Post::transform( get_post( $post ) )->to_object();
+		$activitypub_post = \Activitypub\Transformers_Manager::get_transformer( get_post( $post ) )->transform( get_post( $post ) )->to_object();
 
 		$this->assertEquals( $permalink, $activitypub_post->get_id() );
